refactor: modularize upload status components and unify error constants
- Extract `UploadStatusBar`, `StandardFormatDialog`, and `ExpandableStatusBar` into separate components
- Add `uploadHelpers.tsx` for upload-related utilities
- Simplify `UploadStatusPanel` by composing smaller components
- Introduce `VALIDATION_ERROR` constant and apply to `ValidationError` class

---

refactor: 模組化上傳狀態組件並統一錯誤常數

- 拆分 `UploadStatusBar`、`StandardFormatDialog`、`ExpandableStatusBar` 成獨立元件
- 新增 `uploadHelpers.tsx` 以集中上傳狀態邏輯
- 精簡 `UploadStatusPanel` 結構，提升可維護性
- 新增 `VALIDATION_ERROR` 常數並應用於 `ValidationError` 類
