+++ src/constants.ts
@@ -1,3 +1,5 @@
 export const SUPPORTED_FILE_TYPES = ["application/json", "text/csv"];
 
 export const MAX_FILE_SIZE = 10 * 1024 * 1024; // 10MB
+
+export const VALIDATION_ERROR = "ValidationError";
